[mm] provide helper for finishing mkwrite faults
Message-id: <1507668391-31360-7-git-send-email-lwoodman@redhat.com>
Patchwork-id: 190988
O-Subject: [RHEL7.5 PATCH 06/28] mm: provide helper for finishing mkwrite faults
Bugzilla: 1457569 1383493 1457572
RH-Acked-by: Linda Wang <lwang@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Rik van Riel <riel@redhat.com>

As with the implementation of finish_fault(), we introduce
finish_mkwrite_fault() but we don't refactor the rest of the MM code to use
it.  This is done to reduce code changes in the MM code path to minimize
the chance that we introduce bugs.

This call will only be used by the DAX code, and as such only implements
the wp_pfn_shared() and not the wp_page_shared() version of this
completion as that is all that DAX needs.

Signed-off-by: Ross Zwisler <ross.zwisler@linux.intel.com>

commit 66a6197c118540d454913eef24d68d7491ab5d5f
Author: Jan Kara <jack@suse.cz>
Date:   Wed Dec 14 15:07:39 2016 -0800

    mm: provide helper for finishing mkwrite faults

    Provide a helper function for finishing write faults due to PTE being
    read-only.  The helper will be used by DAX to avoid the need of
    complicating generic MM code with DAX locking specifics.

    Link: http://lkml.kernel.org/r/1479460644-25076-16-git-send-email-jack@suse.cz
    Signed-off-by: Jan Kara <jack@suse.cz>
    Reviewed-by: Ross Zwisler <ross.zwisler@linux.intel.com>
    Acked-by: Kirill A. Shutemov <kirill.shutemov@linux.intel.com>
    Cc: Dan Williams <dan.j.williams@intel.com>
    Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
    Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>

Signed-off-by: Rafael Aquini <aquini@redhat.com>