[target] tcmu: sync up configfs attr setup
Message-id: <1509561736-17598-6-git-send-email-mchristi@redhat.com>
Patchwork-id: 195454
O-Subject: [RHEL 7.5 PATCH 05/10] tcmu: sync up configfs attr setup
Bugzilla: 1507794 1480721 1480434
RH-Acked-by: Ewan Milne <emilne@redhat.com>
RH-Acked-by: Maurizio Lombardi <mlombard@redhat.com>

This is part of upstream commit:

commit 801fc54d5d943e8a6a6bc26bc94fb9b90938ff68
Author: Bryant G. Ly <bryantly@linux.vnet.ibm.com>
Date:   Tue Jun 6 09:28:50 2017 -0500

    tcmu: Make dev_size configurable via userspace

    Allow tcmu backstores to be able to set the device size
    after it has been configured via set attribute.

    Part of support in userspace to support certain backstores
    changing device size.

    Signed-off-by: Bryant G. Ly <bryantly@linux.vnet.ibm.com>
    Reviewed-By: Mike Christie <mchristi@redhat.com>
    Signed-off-by: Nicholas Bellinger <nab@linux-iscsi.org>

Upstream differences:

We are not adding the feature to dynamically change the dev size
in this release, so that part of the upstream patch is dropped.
We are only adding the configfs attr definition cleanup part
of the patch so it matches upstream and it makes the next patches
easier to port.

Signed-off-by: Mike Christie <mchristi@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>