[target] tcmu: make ring buffer timer configurable
Message-id: <1509561736-17598-8-git-send-email-mchristi@redhat.com>
Patchwork-id: 195455
O-Subject: [RHEL 7.5 PATCH 07/10] tcmu: make ring buffer timer configurable
Bugzilla: 1480721 1480434
RH-Acked-by: Ewan Milne <emilne@redhat.com>
RH-Acked-by: Maurizio Lombardi <mlombard@redhat.com>

>From upstream post:

https://www.spinics.net/lists/target-devel/msg16152.html

This adds a timer, qfull_time_out, that controls how long a
device will wait for ring buffer space to open before
failing the commands in the queue. It is useful to separate
this timer from the cmd_time_out and default 30 sec one,
because for HA setups cmd_time_out may be disbled and 30
seconds is too long to wait when some OSs like ESX will
timeout commands after as little as 8 - 15 seconds.

Upstream differences:

The RHEL kernel does not have the upstream dynamic ring buffer
feature so the other changes in that patch are not needed or
are ported to how the RHEL kernel's tcmu module just does a
wait instead of adding to a internal queue.

This patch also removes the pr_err when tcmu_queue_cmd_ring is
called which was done in this patch posted in the same patchset:

https://www.spinics.net/lists/target-devel/msg16170.html

This message is not needed because we will get a log message
from the initiator and/or target already.

Signed-off-by: Mike Christie <mchristi@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>