[target] tcmu: allow max block and global max blocks to be settable
Message-id: <1509561736-17598-11-git-send-email-mchristi@redhat.com>
Patchwork-id: 195460
O-Subject: [RHEL 7.5 PATCH 10/10] tcmu: allow max block and global max blocks to be settable
Bugzilla: 1480721 1480434
RH-Acked-by: Ewan Milne <emilne@redhat.com>
RH-Acked-by: Maurizio Lombardi <mlombard@redhat.com>

>From upstream post:

https://www.spinics.net/lists/target-devel/msg16158.html

Users might have a physical system to a target so they could
have a lot more than 2 gigs of memory they want to devote to
tcmu. On the other hand, we could be running in a vm and so a 2 gig
global and 1 gig per dev limit might be too high. This patch
allows the user to specify the limits.

Upstream differences:

The RHEL kernel does not have the upstream dynamic ring buffer
feature so instead of having code to track/limit pages/blocks
on a per device and global level we only have a preallocated
per device buffer.

Signed-off-by: Mike Christie <mchristi@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>