[pci] Detach driver before procfs & sysfs teardown on device remove
Message-id: <20171109202819.2248.66626.stgit@gimli.home>
Patchwork-id: 197494
O-Subject: [RHEL-7.5 kernel PATCH] PCI: Detach driver before procfs & sysfs teardown on device remove
Bugzilla: 1447196
RH-Acked-by: Jerry Snitselaar <jsnitsel@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1447196
Upstream Status: linux-next via pci queued for v4.15
Build Info: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=14518903
Tested: Attempt to remove via pci-sysfs device in-use and blocked by
        vfio-pci, run lspci while blocked.  Without patch lspci reports
        errors as indicated in upstream commit log.  Repeat with patch,
        lspci remains normal until driver remove callback completes and
        device disappears.

When removing a device, for example a VF being removed due to SR-IOV
teardown, a "soft" hot-unplug via 'echo 1 > remove' in sysfs, or an actual
hot-unplug, we first remove the procfs and sysfs attributes for the device
before attempting to release the device from any driver bound to it.
Unbinding the driver from the device can take time.  The device might need
to write out data or it might be actively in use.  If it's in use by
userspace through a vfio driver, the unbind might block until the user
releases the device.  This leads to a potentially non-trivial amount of
time where the device exists, but we've torn down the interfaces that
userspace uses to examine devices, for instance lspci might generate this
sort of error:

  pcilib: Cannot open /sys/bus/pci/devices/0000:01:0a.3/config
  lspci: Unable to read the standard configuration space header of device 0000:01:0a.3

We don't seem to have any dependence on this teardown ordering in the
kernel, so let's unbind the driver first, which is also more symmetric with
the instantiation of the device in pci_bus_add_device().

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>
Signed-off-by: Bjorn Helgaas <bhelgaas@google.com>
(cherry picked from commit 16b6c8bb687cc3bec914de09061fcb8411951fda)
Signed-off-by: Rafael Aquini <aquini@redhat.com>