[mm] dax: fix DAX PMD pgoff usage
Message-id: <x49inebnxqu.fsf@segfault.boston.devel.redhat.com>
Patchwork-id: 198198
O-Subject: [RHEL7.5 patch] RHEL-only: dax: fix DAX PMD pgoff usage
Bugzilla: 1513605
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>

From: Ross Zwisler

The usage of 'pgoff' in dax_iomap_pmd_fault() is tricky because the local
variable 'pgoff' is the 2MiB aligned page offset, calculated with
'pmd_addr', while vmf->pgoff is the raw page offset calculated with the
actual fault address.

This matters because this check:

    /*
     * Make sure that the faulting address's PMD offset (color) matches
     * the PMD offset from the start of the file.  This is necessary so
     * that a PMD range in the page table overlaps exactly with a PMD
     * range in the radix tree.
     */
    if ((vmf->pgoff & PG_PMD_COLOUR) !=
        ((address >> PAGE_SHIFT) & PG_PMD_COLOUR))
            goto fallback;

really needs to test the raw page offset and not the aligned page offset,
otherwise it will fail for all page faults that aren't 2MiB aligned.  This
will cause us to unnecessarily fail back to 4k faults.

We got this wrong in the initial backport, and were testing with the aligned
'pgoff' variable.

BZ: 1513605
Brew: 14563483 (In Progress)

Signed-off-by: Ross Zwisler <ross.zwisler@linux.intel.com>
Fixes: commit 6031074c9161 ("[fs] dax: fix deadlock due to misaligned PMD faults")
Signed-off-by: Jeff Moyer <jmoyer@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>