[infiniband] ib/core: Add legacy driver's user-data
Message-id: <20171116210913.11180-330-ddutile@redhat.com>
Patchwork-id: 198683
O-Subject: [PATCH RHEL-7.5 329/372] IB/core: Add legacy driver's user-data
Bugzilla: 1499364
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jerry Snitselaar <jsnitsel@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

commit d70724f149b107f8e4062320270d3d8b6713a1bb
Author: Matan Barak <matanb@mellanox.com>
Date:   Thu Aug 3 16:07:04 2017 +0300

    IB/core: Add legacy driver's user-data

    In this phase, we don't want to change all the drivers to use
    flexible driver's specific attributes. Therefore, we add two default
    attributes: UHW_IN and UHW_OUT. These attributes are optional in some
    methods and they encode the driver specific command data. We add
    a function that extract this data and creates the legacy udata over
    it.

    Driver's data should start from UVERBS_UDATA_DRIVER_DATA_FLAG. This

    turns on the first bit of the namespace, indicating this attribute
    belongs to the driver's namespace.

    Signed-off-by: Matan Barak <matanb@mellanox.com>
    Reviewed-by: Yishai Hadas <yishaih@mellanox.com>
    Signed-off-by: Doug Ledford <dledford@redhat.com>

RHEL-7 NOTE:
This patch creates two compile warnings, which result in build errors.
The first is corrected as a change to this patch: init inbuf and outbuf to NULL.
Doug Ledford stated there was a lengthy debate about this issue upstream,
which resolved that this commit is proper as is.
The second error is fixed in the next patch that uses create_udata();
in RHEL, an unused functions generates a warning that becomes a build error;
this warning is fixed in the next patch that uses this function.

Signed-off-by: Donald Dutile <ddutile@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>