[netdrv] bnxt_en: Improve tx ring reservation logic
Message-id: <5243d185ceae3898d560d1153395bfc55c5990f2.1510964611.git.jtoppins@redhat.com>
Patchwork-id: 199157
O-Subject: [RHEL7.5 068/105] bnxt_en: Improve tx ring reservation logic.
Bugzilla: 1459659
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

When the number of TX rings is changed (e.g. ethtool -L, enabling XDP TX
rings, etc), the current code tries to reserve the new number of TX rings
before closing and re-opening the NIC.  If we are unable to reserve the
new TX rings, we abort the operation and keep the current TX rings.

The problem is that the firmware will disable the current TX rings even
when it cannot reserve the new set of TX rings.  We fix it as follows:

1. Instead of reserving the new set of TX rings, just ask the firmware
to check if the new set of TX rings is available.  There is a flag in
the firmware message to do that.  If not available, abort and the
current TX rings will not be disabled.

2. Do the actual TX ring reservation in the path that opens the NIC.
We keep the number of TX rings currently successfully reserved.  If the
number of TX rings is different than the reserved TX rings, we call
firmware and reserve again.

Signed-off-by: Michael Chan <michael.chan@broadcom.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 98fdbe73bfb809b1f8eec9f27a36e737caed3a44)
Bugzilla: 1459659
Build Info: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=14573628
Tested: basic testing further testing being done by Broadcom and myself
Signed-off-by: Jonathan Toppins <jtoppins@redhat.com>

Signed-off-by: Rafael Aquini <aquini@redhat.com>