[kernel] kbuild: Fix tar-pkg with relative $(objtree)
Message-id: <20171128152050.28822-1-dzickus@redhat.com>
Patchwork-id: 200418
O-Subject: [RHEL7 PATCH] kbuild: Fix tar-pkg with relative $(objtree)
Bugzilla: 1518298
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Jonathan Toppins <jtoppins@redhat.com>

Bugzilla: http://bugzilla.redhat.com/1518298

As part of the kernel-ci effort, the project makes
extensive use of the 'make tar-pkg' command to quickly
package and test kernels.   This command does not
work correctly under RHEL-7 with relative paths.

The fix below has zero impact on RHEL kernels as
the tar-pkg command is not used for normal workflow.

It is being requested as part of an effort to build,
package and test RHEL-7 kernels under the kernel-ci
workflow to validate and test ideas.

commit c7eb3a7a1790e0a3a063b4183894d0d63ffee431
Author: Michal Marek <mmarek@suse.cz>
Date:   Wed Jun 18 16:21:41 2014 +0200

    kbuild: Fix tar-pkg with relative $(objtree)

    Commit 7e1c0477 (kbuild: Use relative path for $(objtree)) assumes that
    the build process does not change its working directory. make tar-pkg
    was a couterexample, fix this by changing directory only for the tar
    command and not for the whole script, which at one point references the
    now relative $(objtree).

    Reported-and-tested-by: "J. Bruce Fields" <bfields@fieldses.org>
    Signed-off-by: Michal Marek <mmarek@suse.cz>

Signed-off-by: Rafael Aquini <aquini@redhat.com>