[kernel] wait: fix bit_waitqueue() to allow the use of vmalloc'd memory
Message-id: <1511913420-8761-1-git-send-email-snitzer@redhat.com>
Patchwork-id: 200456
O-Subject: [RHEL-7.5 v2 PATCH 16/14] wait: fix bit_waitqueue() to allow the use of vmalloc'd memory
Bugzilla: 1511159
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Mikulas Patocka <mpatocka@redhat.com>

BZ: 1511159

This is RHEL7-specific and serves as a stop-gap fix that allows the use
of vmalloc'd memory to be used with the kernel's wait code.

Otherwise it is possible to hit a bit_waitqueue() crash, via
wake_up_bit(), if/when memory for the bit field/flags passed to
bit_waitqueue() is vmalloc'd.

With this fix it becomes possible to backport other kv[mz]alloc
consuming upstream commits (e.g. commit 856eb091 "dm: allocate struct
mapped_device with kvzalloc").

In the future we should do a proper RHEL7 backport of upstream commit
9dcb8b68 ("mm: remove per-zone hashtable of bitlock waitqueues").

Suggested-by: Mikulas Patocka <mpatocka@redhat.com>
Signed-off-by: Mike Snitzer <snitzer@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>