[x86] revert "fix get_user_pages() vs device-dax pud mappings"
Message-id: <20171204185916.386-2-lcapitulino@redhat.com>
Patchwork-id: 201030
O-Subject: [RHEL7.5 PATCH 1/2] Revert "[x86] fix get_user_pages() vs device-dax pud mappings"
Bugzilla: 1516984
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Clark Williams <williams@redhat.com>
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>
RH-Acked-by: Daniel Bristot de Oliveira <bristot@redhat.com>

From: Ross Zwisler <ross.zwisler@linux.intel.com>

This reverts commit 6a6ff2ca3ce71a14b229c73d3245c4cc434d1477.

This was a broken backport for the following reasons:

1) It assumes that pud_page(pud) and pte_page(pte) will give the same
value.  This is incorrect - here is some debug from my system:

  gup_huge_pud: pte_page(pte):ffffd80386000000 pud_page(pud):01ffd80386000000

This difference was causing the kernel panic reported here:

https://bugzilla.redhat.com/show_bug.cgi?id=1516984

2) It removed a bunch of RHEL-specific code in the !pud_devmap() case.
This code is present in the !pmd_devmap() case in gup_huge_pmd(), and
should stay.

This commit reverts this backport, and in the next commit we'll do a
smaller backport that will leave the RHEL-only code in place and will
continue to use pte_val(), pte_pfn(), pte_flags() and pte_page() for the
!pud_devmap() case.  This will keep gup_huge_pud() in sync with
gup_huge_pmd() in the RHEL kernel.

I was able to reproduce the kernel panic reported in the bugzilla link
above, and verified that it was fixed with this new backport.  These
patches apply cleanly to kernel-3.10.0-805.el7.

Signed-off-by: Ross Zwisler <ross.zwisler@linux.intel.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>