[infiniband] bnxt_re: fix a crash in qp error event processing
Message-id: <32db19e0322e1520575efea0c89d6f1fb368212e.1512487445.git.jtoppins@redhat.com>
Patchwork-id: 201126
O-Subject: [RHEL7.5 3/6] bnxt_re: fix a crash in qp error event processing
Bugzilla: 1515709
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Jerry Snitselaar <jsnitsel@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

In bnxt_qplib_process_qp_event(), for qp error events we look up the
qp-handle and pass it for further processing. But we don't check if the
handle is NULL. This could lead to a crash in the called functions when
that qp-handle is dereferenced, if the qp is destroyed in the meantime.
Fix this by checking for a valid qp-handle in that function.

Signed-off-by: Sriharsha Basavapatna <sriharsha.basavapatna@broadcom.com>
Signed-off-by: Doug Ledford <dledford@redhat.com>
(cherry picked from commit d6d5c59905c8af932c1cee874e1fb5cd9e83fa61)

Bugzilla: 1515709
Build Info: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=14656450
Tested: ran basic sanity and posted for Broadcom to do additional testing to verify fixes

Signed-off-by: Jonathan Toppins <jtoppins@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>