[md] always hold reconfig_mutex when calling mddev_suspend()
Message-id: <1512741450-26038-14-git-send-email-ncroxon@redhat.com>
Patchwork-id: 202077
O-Subject: [RHEL7.5 PATCH 13/20] md: always hold reconfig_mutex when calling mddev_suspend()
Bugzilla: 1506338
RH-Acked-by: Heinz Mauelshagen <heinzm@redhat.com>
RH-Acked-by: Jerry Snitselaar <jsnitsel@redhat.com>
RH-Acked-by: Xiao Ni <xni@redhat.com>

Conflict:
A minor merge conflict with RHELs version vs upstream because
RHEL7 MD code did not back port this commit 70d466f76.
-Nigel Croxon

commit 4d5324f760aacaefeb721b172aa14bf66045c332
Author: NeilBrown <neilb@suse.com>
Date:   Thu Oct 19 12:17:16 2017 +1100

md: always hold reconfig_mutex when calling mddev_suspend()

Most often mddev_suspend() is called with
reconfig_mutex held.  Make this a requirement in
preparation a subsequent patch.  Also require
reconfig_mutex to be held for mddev_resume(),
partly for symmetry and partly to guarantee
no races with incr/decr of mddev->suspend.

Taking the mutex in r5c_disable_writeback_async() is
a little tricky as this is called from a work queue
via log->disable_writeback_work, and flush_work()
is called on that while holding ->reconfig_mutex.
If the work item hasn't run before flush_work()
is called, the work function will not be able to
get the mutex.

So we use mddev_trylock() inside the wait_event() call, and have that
abort when conf->log is set to NULL, which happens before
flush_work() is called.
We wait in mddev->sb_wait and ensure this is woken
when any of the conditions change.  This requires
waking mddev->sb_wait in mddev_unlock().  This is only
like to trigger extra wake_ups of threads that needn't
be woken when metadata is being written, and that
doesn't happen often enough that the cost would be
noticeable.

Signed-off-by: NeilBrown <neilb@suse.com>
Signed-off-by: Shaohua Li <shli@fb.com>
(cherry picked from commit 4d5324f760aacaefeb721b172aa14bf66045c332)
Signed-off-by: Rafael Aquini <aquini@redhat.com>

Conflicts:
	drivers/md/raid5-cache.c