[scsi] qedf: Limit number of CQs
Message-id: <1513344432-22683-1-git-send-email-cdupuis@redhat.com>
Patchwork-id: 202575
O-Subject: [RHEL 7.5 e-stor PATCH] scsi: qedf: Limit number of CQs
Bugzilla: 1525916
RH-Acked-by: Ewan Milne <emilne@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: Thomas Bogendoerfer <tbogendoerfer@suse.de>

Commit 722477c4f2c50bc8d800d293a3bc5aa843f16e8d

Tested this at Cavium with FCoE and verified that the
number of completion queues is being retrieved from the hardware
allocations done by the qed module and not hard coded.

This will fix the issue Dell is seeing as well as the signature from the
kernel messages matches what is listed below.

FCOE offloading failed with:

[qed_sp_fcoe_func_start:150(sp-0-3b:00.02)]Cannot satisfy CQ amount. CQs
		 requested 8, CQs available 6. Aborting function start
[qed_fcoe_start:821()]Failed to start fcoe
[__qedf_probe:3041]:6: Cannot start FCoE function.

The reason is a newly introduced check in the qed main part. This change
also provides the information about how many CQs are available, so we
simply limit the number of requested CQs..

Fixes: 3c5da9427802 ("qed: Share additional information with qedf")
Signed-off-by: Thomas Bogendoerfer <tbogendoerfer@suse.de>
Reviewed-by: Johannes Thumshirn <jthumshirn@suse.de>
Acked-by: Chad Dupuis <chad.dupuis@cavium.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
Signed-off-by: Bruno E. O. Meneguele <brdeoliv@redhat.com>