[x86] mm/kaiser: introduce user-mapped per-cpu areas
Message-id: <20180104003814.16314-12-aarcange@redhat.com>
Patchwork-id: 203032
O-Subject: [RHEL7.5 PATCH 011/136] x86/mm/kaiser: introduce user-mapped per-cpu areas
Bugzilla: 1519801 1519798 1519786
CVE: CVE-2017-5715 CVE-2017-5753 CVE-2017-5754
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Waiman Long <longman@redhat.com>

From: Dave Hansen <dave.hansen@intel.com>

These patches are based on work from a team at Graz University of
Technology posted here: https://github.com/IAIK/KAISER

The KAISER approach keeps two copies of the page tables: one for running
in the kernel and one for running userspace.  But, there are a few
structures that are needed for switching in and out of the kernel and
a good subset of *those* are per-cpu data.

This patch creates a new kind of per-cpu data that is mapped and
can be used no matter which copy of the page tables is active.
Users of this new section will be forthcoming.

Thanks to Hugh Dickins for cleanups to this code.

Signed-off-by: Dave Hansen <dave.hansen@linux.intel.com>
Cc: Moritz Lipp <moritz.lipp@iaik.tugraz.at>
Cc: Daniel Gruss <daniel.gruss@iaik.tugraz.at>
Cc: Michael Schwarz <michael.schwarz@iaik.tugraz.at>
Cc: Richard Fellner <richard.fellner@student.tugraz.at>
Cc: Andy Lutomirski <luto@kernel.org>
Cc: Linus Torvalds <torvalds@linux-foundation.org>
Cc: Kees Cook <keescook@google.com>
Cc: Hugh Dickins <hughd@google.com>
Cc: x86@kernel.org
Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>