[x86] entry: Move SYSENTER_stack to the beginning of struct tss_struct
Message-id: <20180104003814.16314-41-aarcange@redhat.com>
Patchwork-id: 203045
O-Subject: [RHEL7.5 PATCH 040/136] x86/entry: Move SYSENTER_stack to the beginning of struct tss_struct
Bugzilla: 1519801 1519798 1519786
CVE: CVE-2017-5715 CVE-2017-5753 CVE-2017-5754
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Waiman Long <longman@redhat.com>

From: Andy Lutomirski <luto@kernel.org>

I want SYSENTER_stack to have reliable overflow detection, which
means that it needs to be at the bottom of a page, not the top.
Move it to the beginning of struct tss_struct and page-align it.

Also add an assertion to make sure that the fixed hardware TSS
doesn't cross a page boundary.

Signed-off-by: Andy Lutomirski <luto@kernel.org>
Reviewed-by: Thomas Gleixner <tglx@linutronix.de>
Cc: Borislav Petkov <bp@alien8.de>
Cc: Borislav Petkov <bpetkov@suse.de>
Cc: Brian Gerst <brgerst@gmail.com>
Cc: Dave Hansen <dave.hansen@intel.com>
Cc: Denys Vlasenko <dvlasenk@redhat.com>
Cc: H. Peter Anvin <hpa@zytor.com>
Cc: Josh Poimboeuf <jpoimboe@redhat.com>
Cc: Linus Torvalds <torvalds@linux-foundation.org>
Cc: Peter Zijlstra <peterz@infradead.org>
Link: http://lkml.kernel.org/r/a2dcf4d3c4da0052d1a6c8cabb9f9be257a76b4a.1511195781.git.luto@kernel.org
Signed-off-by: Ingo Molnar <mingo@kernel.org>
Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>