[x86] spec_ctrl: allow the IBP disable feature to be toggled at runtime
Message-id: <20180104003814.16314-119-aarcange@redhat.com>
Patchwork-id: 203162
O-Subject: [RHEL7.5 PATCH 118/136] x86/spec_ctrl: allow the IBP disable feature to be toggled at runtime
Bugzilla: 1519801 1519798 1519786
CVE: CVE-2017-5715 CVE-2017-5753 CVE-2017-5754
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Waiman Long <longman@redhat.com>

On older CPU families the IBP can be turned off (similar to IBRS) but
there's no barrier support like IBPB.

This commit first of all makes sure to initialize the MSR_F15H_IC_CFG
on all CPUs which wasn't happening because of the c != &boot_cpu_data
check (the spec_ctrl_init isn't really supposed to be run by any CPU
but the first one, with the only exception of a post-microcode update
cpuid rescan, but still strictly under mutex).

Secondly this simulates the IBP disabled behavior with ibrs_enabled ==
2 and ibpb_enabled = 1 in debugfs.

The IBP can be re-enabled at runtime with:

    echo 0 > /sys/kernel/debug/x86/ibrs_enabled

and disabled again with:

    echo 2 > /sys/kernel/debug/x86/ibrs_enabled

Setting ibrs_enabled to 1 isn't possible when IBS disable is in use
and it'll return -EINVAL to userland.

It is enforced as invariant that use_ibs_enabled can only be set by
the boot CPU during boot. It also enforced it cannot ever be cleared.

The default ibrs_enabled boot time value if use_ibs_disable is true,
is controlled by USE_IBP_DISABLE_DEFAULT, which if set to 1 will
disable the IBP at boot for maximum security. It's simple to change
the default by setting USE_IBP_DISABLE_DEFAULT to 0 (in which case
it'll be up to the user to echo 2 >ibrs_enabled).

If passing "noibrs noibpb" to the kernel the IBP disable feature won't
be detected and the kernel will just print "FEATURE SPEC_CTRL Not
Present" "FEATURE IBPB_SUPPORT Not Present" signaling the kernel is
running insecure and ibrs_enabled and ibpb_enabled in debugfs will be
forced to 0 and read 0.

If not passing "noibrs noibpb" to the kernel, if IBP disable is
possible the kernel will print "FEATURE SPEC_CTRL Present (Implicit)"
"FEATURE IBPB_SUPPORT Present (Implicit)" to signal ibrs_enabled is
functional and can be set to 2 or 0.

When toggling ibrs_enabled the ibpb_enabled will be updated
accordingly and it'll read 1 when ibrs_enabled is 2, or 0 when
ibrs_enabled is 0, but this purely aesthetical. The important thing
the code checks if use_ibp_disable is true is only !!__ibrs_enabled.

Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>