[x86] spec_ctrl: introduce SPEC_CTRL_PCP_ONLY_IBPB
Message-id: <20180104003814.16314-123-aarcange@redhat.com>
Patchwork-id: 203149
O-Subject: [RHEL7.5 PATCH 122/136] x86/spec_ctrl: introduce SPEC_CTRL_PCP_ONLY_IBPB
Bugzilla: 1519801 1519798 1519786
CVE: CVE-2017-5715 CVE-2017-5753 CVE-2017-5754
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Waiman Long <longman@redhat.com>

This prepares the code to handle setting the SPEC_CTRL_PCP_ONLY_IBPB
mode. This is a special mode where the kernel isn't protected by
(hyper threading/simultaneous multi threading) effects of other CPUs,
but it's fully protected from the userland mode or guest-mode running
in the same thread context. So if this model is combined with
disabling SMT/HT the same security is provided through this model as
if the CPU had the ability to disable or restrict the IBP (for example
with SPEC_CTRL MSR set to IBRS).

Still only in the context of the current thread, this also protects
userland from guest mode, so from that point of view it's equivalent
to ibrs_enabled = 2 (the difference is that ibrs_enabled = 2 also
protects against HT/SMT effects).

This gets enabled by ibpb_enabled = 2 and forces ibrs_enabled set to 0.

Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>