[x86] spec_ctrl: introduce ibpb_enabled = 2 for IBPB instead of IBRS
Message-id: <20180104003814.16314-124-aarcange@redhat.com>
Patchwork-id: 203151
O-Subject: [RHEL7.5 PATCH 123/136] x86/spec_ctrl: introduce ibpb_enabled = 2 for IBPB instead of IBRS
Bugzilla: 1519801 1519798 1519786
CVE: CVE-2017-5715 CVE-2017-5753 CVE-2017-5754
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Waiman Long <longman@redhat.com>

This arms and disarms the SPEC_CTRL_PCP_IBPB_ONLY entry points.

It's automatically enabled at boot if SPEC_CTRL is not available but
IPBP_SUPPORT is available. It's easy to change this not to be
automatically enabled at boot by setting IBPB_ENABLED_ONLY_DEFAULT
from 1 to 0.

It can also be enabled if IBRS is present with:

   echo 2 >/sys/kernel/debug/x86/ibpb_enabled

In which case ibrs_enabled will be forced to 0 and read 0 until
ibpb_enabled is set to a value < 2.

Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>