[x86] kaiser/mm: skip IBRS/CR3 restore when paranoid exception returns to userland
Message-id: <20180104003814.16314-131-aarcange@redhat.com>
Patchwork-id: 203141
O-Subject: [RHEL7.5 PATCH 130/136] x86/kaiser/mm: skip IBRS/CR3 restore when paranoid exception returns to userland
Bugzilla: 1519801 1519798 1519786
CVE: CVE-2017-5715 CVE-2017-5753 CVE-2017-5754
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Waiman Long <longman@redhat.com>

If we schedule while we're in a paranoid exception, we effectively
save the IBRS/CR3 state in the task's kernel stack (they're in
r14/r13d which are callee/schedule saved regs in the task's kernel
stack).

That state is only valid for the exception itself and restoring it
later even potentially in a different CPU, after having gone back to
userland and returned to kernel space would be a bug.

ibrs_enabled may have transitioned from 1 to 2 in the meanwhile just
to make an example and it'd be wrong to clear IBRS then.

Similar considerations for CR3 in r14. pti_enabled may have
transitioned from 0 to 1. The saved cr3 in r14 is the kernel pgd, and
when the task gets scheduled again, it'll return to userland with the
kernel pgd instead of the shadow pgd triggering a segfault with NX set
in the userland part of the kernel pgd.

If we would have needed to restore that specific state and the
userland state would not have been invariant provided by the pcp flags
alone, we could not have scheduled in the first place. So this
simplifies the code and fixes the cr3 NX segfaulting bug at the same
time.

The saving of CR3/IBRS in save_paranoid is all about doing the right
thing when the exception returns to kernel space and we won't schedule
in such case, so all works correctly. Good thing we don't need to
restore the CR3/IBRS state when returning to userland and we can
safely schedule before doing so, so the fix is simple.

Acked-by: Waiman Long <longman@redhat.com>
Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>