[s390] add ppa to system call and program check path
Message-id: <1515614176-17289-4-git-send-email-jcm@redhat.com>
Patchwork-id: 203515
O-Subject: [PATCH 3/3] [RHEL7.5 PATCH] s390: add ppa to system call and program check path
Bugzilla: 1532708
RH-Acked-by: Cornelia Huck <cohuck@redhat.com>

Bugzilla:
https://bugzilla.redhat.com/show_bug.cgi?id=1532708

Description:
This patch mitigates against variant 2 ("Spectre") of the Google
Project Zero vulnerabilities disclosed last week. It does this in
a manner similar to the other architectures, by adding a new type
of indirect branch predictor control interface that can be used
by the kernel to selectively disable indirect prediction when
we are in the kernel. This guards against BTB poisoning.

z/Arch implements this by adding a new PPA-15 facility that we
look for and runtime patch in using alternatives on systems with
updated millicode installed. Older systems will function correctly,
but they will not have mitigation against the vulnerability.

On RHEL, default to forcing the mitigation to be enabled.

Upstream Status:
A modified version of this patch will be posted by IBM engineers.

Brew:
https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=14900816

Test:
Boot tested on emulation and on hardware.

Signed-off-by: Martin Schwidefsky <schwidefsky@de.ibm.com>
Signed-off-by: Jon Masters <jcm@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>