[mm] fix collision between DAX PMD and PTEs
Message-id: <x49inc92okk.fsf@segfault.boston.devel.redhat.com>
Patchwork-id: 203556
O-Subject: [RHEL7.5 patch] RHEL-only: mm: fix collision between DAX PMD and PTEs
Bugzilla: 1528957
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

It turns out that the fix provided by the backport for this commit:

commit e2093926a098 ("dax: fix race between colliding PMD & PTE entries")

is insufficient in RHEL because we are missing a backport for this commit:

commit 7267ec008b5c ("mm: postpone page table allocation until we have page to map")

7267ec008b5c changes the time at which the PTE-parent PMD is installed in a
page table in response to a page fault.  Prior to 7267ec008b5c the PMD was
installed via pte_alloc() inside of __handle_mm_fault().  This flow is the
same in RHEL, with the PMD installation happening via __pte_alloc()
instead.

After 7267ec008b5c, the PMD is installed via alloc_set_pte() =>
pte_alloc_one_pte() when the PTE is being installed.

For DAX this distinction is important because with the older code (and
RHEL) this PMD insertion happens outside of our radix tree locking.  With
the alloc_set_pte() path, we are holding a radix tree lock.  This allows us
to re-check the PMD under the radix tree lock, which is what the fix in
e2093926a098 does.

Unfortunately 7267ec008b5c is part of a much larger reorganization of the
memory management code, and is not easily pulled in without adding risk.

Instead we fix this race by verifying our PMD insertion in
insert_pfn_pmd(), under the ptl lock.  If we have a collision we follow the
lead of the equivalent check in dax_pmd_load_hole() and just return
VM_FAULT_FALLBACK, causing the FS DAX fault handlers to fall back to 4k
faults.

This race should never happen for device DAX, but if we do hit this
condition we SIGBUS as the device DAX code can never mix hugepage PMDs and
PTEs in the same namespace.

In my testing I was able to reproduce all variants of this issue:

https://bugzilla.redhat.com/show_bug.cgi?id=1528957

The kernel panic on bare metal and the dmesg warning in mm/mmap.c:3040 in a
kvm guest were both fixed by this patch.

Signed-off-by: Ross Zwisler <ross.zwisler@linux.intel.com>
Signed-off-by: Jeff Moyer <jmoyer@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <brdeoliv@redhat.com>