[nvdimm] btt: fix uninitialized err_lock
Message-id: <x497esi6ddw.fsf@segfault.boston.devel.redhat.com>
Patchwork-id: 203816
O-Subject: [rhel7.5 patch] btt: fix uninitialized err_lock
Bugzilla: 1524775
RH-Acked-by: David Milburn <dmilburn@redhat.com>
RH-Acked-by: Eric Sandeen <esandeen@redhat.com>

BZ: 1524775
Brew: 15006446

Upstream: This patch was posted to the linux-nvdimm mailing list [1] and
acked, but has yet to make it into the nvdimm tree.  We need to fix this
before rhel 7.5 ships, so I'm posting this w/o an upstream commit ID.

When a sector mode namespace is initially created, the arena's err_lock
is not initialized.  If, on the other hand, the namespace already
exists, the mutex is initialized.  To fix the issue, I moved the mutex
initialization into the arena_alloc, which is called by both
discover_arenas and create_arenas.

This was discovered on an older kernel where mutex_trylock checks the
count to determine whether the lock is held.  Because the data structure
is kzalloc-d, that count was 0 (held), and I/O to the device would hang
forever waiting for the lock to be released (see btt_write_pg, for
example).  Current kernels have a different mutex implementation that
checks for a non-null owner, and so this doesn't show up as a problem.
If that lock were ever contended, it might cause issues, but you'd have
to be really unlucky, I think.

Signed-off-by: Jeff Moyer <jmoyer@redhat.com>

[1] https://lists.01.org/pipermail/linux-nvdimm/2017-December/013569.html

Signed-off-by: Bruno E. O. Meneguele <brdeoliv@redhat.com>