[netdrv] ibmvnic: Fix pending MAC address changes
Message-id: <1516235360-15400-1-git-send-email-drosario@redhat.com>
Patchwork-id: 204043
O-Subject: [RHEL-7.5 BZ1535368 PATCH] ibmvnic: Fix pending MAC address changes
Bugzilla: 1535368
RH-Acked-by: Steve Best <sbest@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1535368

DESCRIPTION:
------------
Due to architecture limitations, the IBM VNIC client driver is unable
to perform MAC address changes unless the device has "logged in" to
its backing device. Currently, pending MAC changes are handled before
login, resulting in an error and failure to change the MAC address.
Moving that chunk to the end of the ibmvnic_login function, when we are
sure that it was successful, fixes that.

The MAC address can be changed when the device is up or down, so
only check if the device is in a "PROBED" state before setting the
MAC address.

Fixes: c26eba03e407 ("ibmvnic: Update reset infrastructure to support tunable parameters")
Signed-off-by: Thomas Falcon <tlfalcon@linux.vnet.ibm.com>
Reviewed-by: John Allen <jallen@linux.vnet.ibm.com>
Signed-off-by: David S. Miller <davem@davemloft.net>

UPSTREAM STATUS:
----------------
commit 3d1661304f0b2b51a8a43785b764822611dbdd53

https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=3d1661304f0b2b51a8a43785b764822611dbdd53

BREW:
-----
https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=15033691

TESTING:
--------
Full testing is yet to be performed by IBM since Red Hat does not have this
hardware.

For a sanity test I made sure the drive was able to load successfully.

[root@rhel-7 ibm]# insmod ibmvnic.ko
[root@rhel-7 ibm]# lsmod | grep ibmvnic
ibmvnic                82769  0

[root@rhel-7 ibm]# dmesg | grep 'ibmvnic registered\|NIC'
[    0.836524] ibmvnic: IBM System i/p Virtual NIC Driver 1.0
[    0.857270] ibmvnic 30000005: ibmvnic registered

----------------
Desnes Augusto Nunes do Rosario

Signed-off-by: Bruno E. O. Meneguele <brdeoliv@redhat.com>