[fs] gfs2: Prevent stopping kthreads twice in gfs2_make_fs_ro
Message-id: <20180125172343.16265-1-agruenba@redhat.com>
Patchwork-id: 204395
O-Subject: [RHEL7 fs] gfs2: Prevent stopping kthreads twice in gfs2_make_fs_ro
Bugzilla: 1538344
RH-Acked-by: Andrew Price <anprice@redhat.com>
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>

Bugzilla: 1538344
Upstream status: not needed
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=15104116
Tested: locally

When a filesystem is withdrawn and then remounted read-only,
gfs2_make_fs_ro will stop the filesystem's logd and quotad kthreads and
then fail acquiring the transaction glock.  The filesystem will remain
mounted read-write.  When gfs2_make_fs_ro is called again, as part of
unmounting the filesystem or otherwise, it will try stopping the
kthreads a second time and crash.  Fix that by setting the kthread
pointers to NULL when the kthreads are not running.

In addition, acquiring the transaction glock could fail for other
reasons after the kthreads have been stopped when the filesystem is not
being withdrawn.  In that case, try to keep the filesystem running by
restarting the kthreads. If that isn't possible, our only option is to
withdraw the filesystem.

Finally, it makes no sense to deny remounting a filesystem read-only
when the filesystem is already withdrawn, so don't fail gfs2_make_fs_ro
in that case.

(In mainline, the transaction glock doesn't exist anymore and we are
acquiring the freeze glock before stopping the kthreads, so there is no
need for any of this.)

Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>
Signed-off-by: Bob Peterson <rpeterso@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>